requirements.txt
streamlit
pandas
numpy
matplotlib
folium
streamlit-folium
geopy
psycopg2-binary
python-dotenv
anthropic
